refactor: sebrate form input from createMatch component
