+++ client/package.json
@@ -25,6 +25,7 @@
     "@mui/material": "^5.13.5",
     "@mui/system": "^5.13.5",
     "@mui/types": "^7.2.4",
+    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.9.0",
