+++ client/src/components/createMatchComponents/index.ts
@@ -1,0 +1,3 @@
+import CreateMatchForm from './CreateMatchForm';
+
+export default CreateMatchForm;
