+++ client/src/components/index.ts
@@ -12,6 +12,7 @@ export { default as Definition } from './Definition/Definition';
 export { default as Features } from './Features/Features';
 export { default as BestStadiums } from './best-stadiums/BestStadiums';
 export { default as Contact } from './contact-us/ContactUs';
+// export { default as CreateMatchForm } from './createMatchComponents';
 
 export {
   StyledContainer,
