+++ client/src/context/CreateMatch.tsx
@@ -1,0 +1,88 @@
+import PropTypes from 'prop-types';
+
+import { ReactNode, createContext, useEffect, useMemo, useState } from 'react';
+import { IEvent } from '../pages/CreateMatch';
+import { OpenContextValue2 } from '../interfaces';
+
+interface MyContextProviderProps {
+  children: ReactNode;
+}
+
+const statsContext = createContext<OpenContextValue2 | any>({});
+statsContext.displayName = 'OpenContext';
+
+const OpenContextProvider = ({ children }: MyContextProviderProps) => {
+  const [Stadiums, setStadiums] = useState<any[]>([]);
+  const [UserId, setUserId] = useState<number>(0);
+  const [Details, setDetails] = useState<any>();
+  const [ValidateError, setValidateError] = useState<string>('');
+  const [matches, setMatches] = useState<IEvent[]>([]);
+  const [Event, setEvent] = useState<IEvent>({
+    title: '',
+    start: '',
+    end: '',
+    backgroundColor: '',
+  });
+  const [match, setMatch] = useState({
+    title: '',
+    startDate: '',
+    endDate: '',
+    seats: 0,
+    description: '',
+    UserId: 0,
+  });
+  // const contextValue = useContext(open);
+  // const { openPage, updateOpen } = contextValue;
+
+  // useEffect(() => {
+  //   setMatch({
+  //     title: '',
+  //     startDate: '',
+  //     endDate: '',
+  //     seats: 0,
+  //     description: '',
+  //     UserId: 0,
+  //   });
+  //   setEvent({
+  //     title: '',
+  //     start: '',
+  //     end: '',
+  //     backgroundColor: '',
+  //   });
+  //   setValidateError('');
+  // }, [openPage]);
+
+  useEffect(() => {
+    setMatch(prv => ({ ...prv, startDate: Event.start, endDate: Event.end }));
+  }, [Event]);
+
+  const values = useMemo(
+    () => ({
+      Stadiums,
+      setStadiums,
+      UserId,
+      setUserId,
+      Details,
+      setDetails,
+      ValidateError,
+      setValidateError,
+      matches,
+      setMatches,
+      Event,
+      setEvent,
+      match,
+      setMatch,
+    }),
+    [Stadiums, matches, Event, match, ValidateError, Details, UserId],
+  );
+
+  return (
+    <statsContext.Provider value={values}>{children}</statsContext.Provider>
+  );
+};
+
+export { statsContext, OpenContextProvider };
+
+OpenContextProvider.propTypes = {
+  children: PropTypes.node.isRequired,
+};
