--- client/src/context/createMatcheContext.tsx
@@ -1,11 +1,0 @@
-import { createContext } from 'react';
-
-interface OpenContextValue {
-  openPage?: boolean;
-  setOpenPage?: React.Dispatch<React.SetStateAction<boolean>>;
-}
-
-const open = createContext<OpenContextValue>({});
-open.displayName = 'open';
-
-export default open;
