+++ client/src/context/index.ts
@@ -1,3 +1,1 @@
-import open from './createMatcheContext';
-
-export default open;
+export { open, useCustomOpen } from './RootLayout';
