+++ client/src/interfaces/matchInterface.ts
@@ -1,12 +1,15 @@
 import { Dispatch, SetStateAction } from 'react';
+import { IEvent } from '../pages/CreateMatch';
 
 export interface Option {
   id: number;
   username: string;
 }
+export interface LeftSideBarInterface {
+  setOpen: (newValue: boolean) => void;
+}
 
-export interface createMatchInterface {
-  setOpen: React.Dispatch<React.SetStateAction<boolean>>;
+export interface createMatchInterface extends LeftSideBarInterface {
   open: boolean;
 }
 
@@ -43,4 +46,26 @@ export interface prevInterface {
 export interface newEventInterface {
   startStr: string;
   endStr: string;
+}
+
+export interface OpenContextValue {
+  openPage?: boolean;
+  updateOpen?: (newValue: boolean) => void;
+}
+
+export interface OpenContextValue2 {
+  Stadiums: any[];
+  setStadiums: React.Dispatch<React.SetStateAction<any[]>>;
+  UserId: number;
+  setUserId: React.Dispatch<React.SetStateAction<number>>;
+  Details: any;
+  setDetails: React.Dispatch<React.SetStateAction<any>>;
+  ValidateError: string;
+  setValidateError: React.Dispatch<React.SetStateAction<string>>;
+  matches: IEvent[];
+  setMatches: React.Dispatch<React.SetStateAction<IEvent[]>>;
+  Event: IEvent;
+  setEvent: React.Dispatch<React.SetStateAction<IEvent>>;
+  match: prevInterface;
+  setMatch: React.Dispatch<React.SetStateAction<prevInterface>>;
 }
