+++ client/src/main.tsx
@@ -4,9 +4,12 @@ import ReactDOM from 'react-dom/client';
 import { ThemeProvider } from '@emotion/react';
 import LightTheme from './themes';
 import App from './App';
+import { OpenContextProvider } from './context/CreateMatch';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <ThemeProvider theme={LightTheme}>
-    <App />
+    <OpenContextProvider>
+      <App />
+    </OpenContextProvider>
   </ThemeProvider>,
 );
